KRIPTON-71: Make parser xml/jackson - basic types
Task-Url: http://xsvn:8080/jira/browse/KRIPTON-71

+++ KriptonProcessor/src/test/java/kripton70/BeanElement70.java
@@ -22,7 +22,7 @@ public class BeanElement70 {
 	}
 	
 	@BindXml(xmlType=XmlType.VALUE_CDATA)
-	public short valueContentBoolType;
+	public Integer valueContentBoolType;
 	
 	public boolean valueBoolType;
